==> 5_RelationalClassAnalysis/0_Source.R <==
##### SOURCE FILE FOR "PARTY, RACE, AND NEUTRALITY: INVESTIGATING THE INTERDEPENDENCE OF ATTITUDES TOWARDS SOCIAL GROUPS"
##### BY BRENSINGER & SOTOUDEH


## Step 1: Install packages
source("Brensinger2022/1_Packages.R")

## Step 2: Define functions
source("Brensinger2022/2_Functions.R")

## Step 3: Import and clean data and run RCA
source("Brensinger2022/3_Running_rca.R")

## Step 4: Variable cleaning and creation
source("Brensinger2022/4_Variable_cleaning_and_creation.R")

## Step 5: Scale construction
source("Brensinger2022/5_Scale_construction.R")

## Step 6: Descriptives and regressions
source("Brensinger2022/6_Descriptives_and_regressions.R")

## Step 7: Final black and white plots
source("Brensinger2022/7_Final_bw_plots_tiff.R")

## Step 8: Supplementary analyses
# source("Brensinger2022/8_Supplementary_analyses.R")
